+++ CHANGELOG.md
@@ -20,6 +20,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
   [Uku Loskit](https://github.com/UkuLoskit)
   [#6037](https://github.com/CocoaPods/CocoaPods/issues/6037)
 
+* The validator has an API for accessing which version of Swift was used.    
+  [Orta Therox](https://github.com/orta)
+  [#6049](https://github.com/CocoaPods/CocoaPods/pull/6049)
+
 
 ##### Bug Fixes